Refactoeing components directory
