+++ src/App.tsx
@@ -1,6 +1,6 @@
 import React from "react";
-import Root from "./components/Root";
-import Header from "./components/Header";
+import Root from "components/Root";
+import Header from "components/Header";
 import Section from "components/Section";
 import ContractWallet from "components/ContractWallet";
 import ContractMain from "components/ContractMain";
