+++ src/components/ContractMain/styles.tsx
@@ -1,0 +1,1 @@
+export {};
