+++ src/components/Navigation/index.tsx
@@ -1,0 +1,1 @@
+export {};
