+++ src/components/Root/index.tsx
@@ -1,7 +1,5 @@
 import React, { FC } from "react";
-import { classnames } from "classnames/tailwind";
-
-const rootStyle = classnames("container", "mx-auto", "px-5", "pb-10");
+import { rootStyle } from "./styles";
 
 const Root: FC = ({ children }) => {
   return <main className={rootStyle}>{children}</main>;
