+++ src/components/Root/styles.tsx
@@ -1,0 +1,3 @@
+import { classnames } from "classnames/tailwind";
+
+export const rootStyle = classnames("container", "mx-auto", "px-5", "pb-10");
