+++ src/components/Section/index.tsx
@@ -1,19 +1,9 @@
 import React, { FC } from "react";
-import { classnames } from "classnames/tailwind";
+import { sectionBlockStyle, sectionStyle, sectionTitleStyle } from "./styles";
 
 interface SectionProps {
   title?: string;
 }
-
-const sectionStyle = classnames("mt-4");
-const sectionTitleStyle = classnames(
-  "text-2xl",
-  "uppercase",
-  "font-medium",
-  "text-base-black",
-  "mb-2"
-);
-const sectionBlockStyle = classnames("bg-base-gray", "rounded-base");
 
 const Root: FC<SectionProps> = ({ children, title }) => {
   return (
