+++ src/components/Section/styles.tsx
@@ -1,0 +1,11 @@
+import { classnames } from "classnames/tailwind";
+
+export const sectionStyle = classnames("mt-4");
+export const sectionTitleStyle = classnames(
+  "text-2xl",
+  "uppercase",
+  "font-medium",
+  "text-base-black",
+  "mb-2"
+);
+export const sectionBlockStyle = classnames("bg-base-gray", "rounded-base");
